[userbouquet] optimize and add option to delete

+++ lib/python/Screens/ChannelSelection.py
@@ -124,6 +124,18 @@ def append_when_current_valid(current, menu, args, level = 0, key = ""):
 	if current and current.valid() and level <= config.usage.setup_level.index:
 		menu.append(ChoiceEntryComponent(key, args))
 
+def removed_userbouquets_available():
+	for file in os.listdir("/etc/enigma2/"):
+		if file.startswith("userbouquet") and file.endswith(".del"):
+			return True
+	return False
+
+def removed_userbouquets_not_available():
+	for file in os.listdir("/etc/enigma2/"):
+		if file.startswith("userbouquet") and file.endswith(".del"):
+			return False
+	return True
+
 class ChannelContextMenu(Screen):
 	def __init__(self, session, csel):
 
@@ -222,7 +234,9 @@ class ChannelContextMenu(Screen):
 			if csel.bouquet_mark_edit == OFF:
 				if not csel.movemode:
 					append_when_current_valid(current, menu, (_("enable move mode"), self.toggleMoveMode), level = 1, key = "yellow")
-					append_when_current_valid(current, menu, (_("restore deleted userbouquets"), self.restoreDeletedBouquets), level=0)
+					if removed_userbouquets_available():
+						append_when_current_valid(current, menu, (_("restore deleted userbouquets"), self.restoreDeletedBouquets), level=0)
+						append_when_current_valid(current, menu, (_("purge deleted userbouquets"), self.purgeDeletedBouquets), level=0)
 					if not inBouquetRootList and current_root and not (current_root.flags & eServiceReference.isGroup):
 						if current.type != -1:
 							menu.append(ChoiceEntryComponent(text = (_("add marker"), self.showMarkerInputBox)))
@@ -476,18 +490,33 @@ class ChannelContextMenu(Screen):
 		self.close()
 
 	def restoreDeletedBouquets(self):
-		for file in os.listdir("/etc/enigma2/"):
-			if file.startswith("userbouquet") and file.endswith(".del"):
-				file = "/etc/enigma2/" + file
-				print "restore file ", file[:-4]
-				os.rename(file, file[:-4])
-		eDVBDBInstance = eDVBDB.getInstance()
-		eDVBDBInstance.setLoadUnlinkedUserbouquets(True)
-		eDVBDBInstance.reloadBouquets()
-		#eDVBDBInstance.setLoadUnlinkedUserbouquets(config.misc.load_unlinked_userbouquets.value)
-		refreshServiceList()
-		self.csel.showFavourites()
-		self.close()
+		self.session.openWithCallback(self.restoreDeletedBouquetsCallback, MessageBox, _("Are you sure to restore all deleted userbouquets?"))
+
+	def restoreDeletedBouquetsCallback(self, answer):
+		if answer:
+			for file in os.listdir("/etc/enigma2/"):
+				if file.startswith("userbouquet") and file.endswith(".del"):
+					file = "/etc/enigma2/" + file
+					print "restore file ", file[:-4]
+					os.rename(file, file[:-4])
+			eDVBDBInstance = eDVBDB.getInstance()
+			eDVBDBInstance.setLoadUnlinkedUserbouquets(True)
+			eDVBDBInstance.reloadBouquets()
+			refreshServiceList()
+			self.csel.showFavourites()
+			self.close()
+
+	def purgeDeletedBouquets(self):
+		self.session.openWithCallback(self.purgeDeletedBouquetsCallback, MessageBox, _("Are you sure to purge all deleted userbouquets?"))
+
+	def purgeDeletedBouquetsCallback(self, answer):
+		if answer:
+			for file in os.listdir("/etc/enigma2/"):
+				if file.startswith("userbouquet") and file.endswith(".del"):
+					file = "/etc/enigma2/" + file
+					print "permantly remove file ", file
+					os.remove(file)
+			self.close()
 
 class SelectionEventInfo:
 	def __init__(self):
